fix: constant HelmRelease updates
This PR reworks `ReconcileHelmRelease` function and updates the way
how values are handled. When two values of the apiextensionsv1.JSON
type are compared (expected vs actual values) it's considered as not
equal even if the internal values are the same.

This leads to constant update of the HelmRelease object. It does not
affect components without custom config specified.

Fixes #1398