fix: Rework the helm-push Makefile target
Treat any failed pull (including 403 or other errors) as if
the chart does not exist, and always attempt to push it.

Currently, pull errors like 403 Forbidden are silently ignored
and incorrectly interpreted as if the chart already exists in the repository.
This prevents charts from being pushed when they actually should be.