fix: pass cluster name explicitly to ccm
This fix provides options to avoid OpenStack naming resources conflicts
in the cases where CCM can "see" other projects and confuse their resources
by name with the resources of other clusters. It happens when `cluster.name`
propery is not set and thus a default of "kubernetes" applies. The upstream
chart has it since really old times, so we need to adopt it to avoid falling
into that trap. Unfortunately, there are still bugs in upstream which would
not allow seamless change of this variable. Therefore, users should be advised
to enable an additional flag for the clusters already installed and do the modification
during a maintenance window to avoid service interruption. Otherwise, it should
be completely okay to apply this change for new clusters.