feat: do not prefix regional releases with region name
Previously helm releases were installed with a name that equals
to the HelmRelease object name. With this change the new field
(ReleaseName) was introduced to the ReconcileHelmReleaseOpts
that allows to override the name of the deployed helm release.
It's needed to avoid unnecessary prefix for all the regional
helm releases.