fix: add regional section to the Release spec
With the old approach, the kcm-regional version depended on the kcm version.
During upgrades, the previous release was marked as not ready because the kcm
controller attempted to find a template named `kcm-regional-<old-version>`,
which did not exist.

To resolve this, `spec.regional` is reintroduced in the Release object, but
`spec.regional.template` remains unset in v1.4.0.
This makes the upgrade path from v1.3.1 → v1.4.0 possible.
After v1.4.0 is released, this can be reworked.

Partially reverts https://github.com/k0rdent/kcm/pull/1996

Related to #1903